Time: 4 ms (97.40%), Space: 17.3 MB (48.23%) - LeetHub

+++ 662-maximum-width-of-binary-tree/662-maximum-width-of-binary-tree.cpp
@@ -5,22 +5,22 @@ public:
         if (!root) return 0;
         q.push({root, 0});
         
-        long long width = 1; 
+        int width = 1; 
         while (!q.empty()) {
             int size = q.size();
-            long long minVertical = q.front().second, maxVertical = q.back().second;
+            int minVertical = q.front().second, maxVertical = q.back().second;
             
             for (auto i = 0; i < size; i++) {
                 TreeNode *front = q.front().first;
-                long long idx = q.front().second - minVertical;    
+                int idx = q.front().second - minVertical;    
                 
-                if (front->left) q.push({front->left, 2 * idx + 1}); 
-                if (front->right) q.push({front->right, 2 * idx + 2}); 
+                if (front->left) q.push({front->left, (long long)2 * idx + 1}); 
+                if (front->right) q.push({front->right, (long long)2 * idx + 2}); 
                 q.pop();
             };
             width = max(width, maxVertical - minVertical + 1);
         };
-        return (int)width;
+        return width;
     };
 };
